Move entity to entities folder instead of root of module
