+++ src/database/seeders/posts/posts.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { PostToFlair } from "../../../posts/postToFlair.entity";
+import { PostToFlair } from "../../../posts/entities/postToFlair.entity";
 import { Post } from "../../../posts/entities/post.entity";
 import { PostsSeederService } from "./posts.service";
